backend/requirements.txt
# Flask Core
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Bcrypt==1.0.1

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
GeoAlchemy2==0.14.2

# Validation
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
email-validator==2.1.0

# ML/NLP
scikit-learn==1.3.2
nltk==3.8.1
numpy==1.26.2
pandas==2.1.3
joblib==1.3.2

# File handling
Werkzeug==3.0.1
Pillow==10.1.0
python-magic-bin==0.4.14

# Utils
python-dotenv==1.0.0
gunicorn==21.2.0
celery==5.3.4
redis==5.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
factory-boy==3.3.0
